+++ Search/AdapterInterface.php
@@ -1,6 +1,16 @@
 <?php
+/*
+ * This file is part of the Sulu CMS.
+ *
+ * (c) MASSIVE ART WebServices GmbH
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
 
 namespace Massive\Bundle\SearchBundle\Search;
+
+use Massive\Bundle\SearchBundle\Search\SearchQuery;
 
 /**
  * Interface to be implement by all search library adapters
@@ -10,7 +20,7 @@ namespace Massive\Bundle\SearchBundle\Search;
 interface AdapterInterface
 {
     /**
-     * Index the given IndexEntry object
+     * Index the given Document object
      *
      * @param Document $document Document to index
      * @param string $indexName Name of index to store document in
@@ -18,11 +28,19 @@ interface AdapterInterface
     public function index(Document $document, $indexName);
 
     /**
+     * Remove the given Document from the index
+     *
+     * @param Document $document
+     * @param string $indexName
+     */
+    public function deindex(Document $document, $indexName);
+
+    /**
      * Search using the given query string
      *
      * @param string $queryString
      */
-    public function search($queryString, array $indexNames = array());
+    public function search(SearchQuery $searchQuery);
 
     /**
      * Return vendor status information as an associative